Update for week3 homework
Please upload the modification you made to the code which prompted the
user for the events to search for in your menu script.

Additionally, modify the menu script that checks if the user specified
multiple search strings.

If so, modify the search string input so that it includes the
parenthesis and replace the commas with a pipe sign => | and then
perform the search.